feat: remove truststore=None alias of truststore=TrustStore.system()
Instantiating a TLSServerConfiguration with `truststore=None` means that
client-authentication is disabled. Instantiating a TLSClientConfiguration
with the same `truststore=None` means instead to use the system trust
store. Instantiating a TLSClientConfiguration with
`truststore=TrustStore.system()` is another (more explicit) way to use
the system trust store.

We believe that there should only be one way to instantiate a client
configuration so it uses the system trust store, and that `None` is a
bad value as it has a totally different meaning in the server
configuration.

We used the opportunity of this work to introduce `is_system()`: a new
property on `TrustStore` that should be called to test if a given
truststore object is the system one. This way the implementation (all
three attributes `_buffer`, `_path` and `_id` set `None`) becomes an
implementation detail.